Version 1.0.9 (Follower OnDeath issue fixed)
The follower is now prevented from making a Lost Leader check if they
are dead. Without the check on the follower being dead, the Lost Leader
call resulted in the follower getting stuck in a perpetual loop, as it
could not TP back to town.
Pre-release changes/additions:
1.0.8.3 - added Crystal Colonnade and Leroric's Passage to Boss
Encounter code
1.0.8.4 - OnBotStart and OnEnabled now initialise all the variables,
thus allowing the bots to be stopped and restrted during a run without
having to reform the party.