Version 2.0.4 (stuck on Azmodan portal fixed)
And this time it really has been fixed!
A check is made for when the follower is within 100 of Amzodan's portal
and disables portal usage. decision is based on the LocalAreaID and the
coordinates of the potal itself.
It is left to the leader to activate Azmodan's portal. The followers
will still accept the Boss Encounter invite.